services/backend/requirements.txt
flask>=3.0
flask-cors>=4.0
psycopg2-binary>=2.9
python-dotenv>=1.0
azure-storage-blob>=12.19
gunicorn>=21.0
